organized package code (resolves #9) and added pre-commit (closes #10)
