+++ .github/workflows/publish.yml
@@ -16,14 +16,14 @@ jobs:
 
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v1
-            
+
       - name: Log in to the GitHub Container Registry
         uses: docker/login-action@v1
         with:
           registry: ghcr.io
           username: ${{ github.actor }}
           password: ${{ secrets.GITHUB_TOKEN }}
-          
+
       - name: Build and push to GitHub Container Registry
         uses: docker/build-push-action@v2
         with:
