+++ LICENSE.txt
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2022 Diogo Correia
+Copyright (c) 2022 seizmeia dev team: Diogo Correia, Simão Matos
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
