+++ requirements-dev.txt
@@ -1,9 +1,10 @@
 -r requirements.txt
-pytest~=6.2.5
-mypy~=0.931
-coverage~=6.3.1
-black~=22.1.0
-isort~=5.10.1
-tox~=3.24.5
-types-PyYAML~=6.0.4
-pytest-asyncio~=0.18.1
+black>=22.1.0
+coverage>=6.3.1
+isort>=5.10.1
+mypy>=0.931
+pre-commit>=2.17.0
+pytest>=6.2.5
+pytest-asyncio>=0.18.1
+tox>=3.24.5
+types-PyYAML>=6.0.4
