+++ requirements.txt
@@ -1,5 +1,5 @@
-uvicorn~=0.17.4
 fastapi==0.73.0
-python-multipart==0.0.5
 passlib==1.7.4
+python-multipart==0.0.5
 PyYAML==5.4.1
+uvicorn~=0.17.4
