+++ seizmeia/auth.py
@@ -1,5 +1,7 @@
-from enum import Enum
+from __future__ import annotations
+
 from datetime import timedelta
+from enum import Enum
 
 from pydantic import BaseModel
