+++ setup.cfg
@@ -1,30 +1,41 @@
 [metadata]
-name         = seizmeia
-version      = 0.0.1
-author       = Diogo Correia, Simão Matos
-author_email = dv_correia@hotmail.com, simaopedromats@gmail.com
-license      = MIT
-platforms    = any
-description  = A credit management tool for a beer tap.
+name = seizmeia
+version = 0.0.1
+description = A credit management tool for a beer tap.
 long_description = file: README.md
 long_description_content_type = text/markdown
-keywords     = beer, tap, management
 url = https://github.com/seizmeia/seizmeia
-project_urls =
-    Source Code = https://github.com/seizmeia/seizmeia
-    Bug Reports = https://github.com/seizmeia/seizmeia/issues
+author = Diogo Correia, Simão Matos
+author_email = dv_correia@hotmail.com, simaopedromats@gmail.com
+license = MIT
+license_file = LICENSE.txt
+platforms = any
+classifiers =
+    License :: OSI Approved :: MIT License
+    Programming Language :: Python :: 3
+    Programming Language :: Python :: 3 :: Only
+    Programming Language :: Python :: 3.10
 classifier =
     Development Status :: 1 - Planning
     Intended Audience :: End Users/Desktop
     License :: OSI Approved :: MIT License
     Programming Language :: Python :: 3.10
+keywords = beer, tap, management
+project_urls =
+    Source Code = https://github.com/seizmeia/seizmeia
+    Bug Reports = https://github.com/seizmeia/seizmeia/issues
 
 [options]
 packages = find:
-    
+python_requires = >=3.10
+
+[options.packages.find]
+exclude =
+    tests*
+
 [options.entry_points]
 console_scripts =
-    seizmeia = seizmeia:run
+    seizmeia = seizmeia.__main__:run
 
 [tool:pytest]
 asyncio_mode = strict
