+++ tests/__init__.py
@@ -1,0 +1,1 @@
+# This page intentionally left blank
