+++ tox.ini
@@ -1,6 +1,6 @@
 [tox]
 minversion = 3.3.0
-envlist = 
+envlist =
     python3.10,
     lint,
     type,
@@ -8,24 +8,30 @@ isolated_build = true
 
 [testenv]
 description = run the test driver with {basepython}
-deps = 
-    pytest
-    -r{toxinidir}/requirements-dev.txt
-commands = python -m pytest {posargs}
+deps = -r{toxinidir}/requirements-dev.txt
+commands = 
+    coverage erase
+    coverage run -m pytest {posargs:tests}
+    coverage report
 
 [testenv:lint]
 description = check the code style
-deps = 
+deps =
     flake8
     black
-commands = 
+commands =
     python -m black --line-length 79 --diff --color --check seizmeia
     python -m flake8 {posargs}
 
 [testenv:type]
 description = type check
-deps = 
+deps =
     mypy
     -r{toxinidir}/requirements-dev.txt
 commands =
     mypy -p seizmeia
+
+[testenv:pre-commit]
+skip_install = true
+deps = pre-commit
+commands = pre-commit run --all-files --show-diff-on-failure
